refactor: 루틴 api 리팩토링
