+++ src/domains/middlewares/validateJournalIdQuery.js
@@ -1,8 +1,8 @@
 import { isUUID } from '../utils/validator.utils.js';
 
-export const validateJournalId = (req, res, next) => {
+export const validateJournalIdQuery = (req, res, next) => {
   try {
-    const { journalId } = req.params;
+    const { journalId } = req.query;
     if (!isUUID(journalId)) {
       throw new Error('일지 ID는 유효한 UUID 형식이 아닙니다.');
     }
